mario lib for n64 compilation

+++ src/engine/math_util.c
@@ -1117,7 +1117,7 @@ static u32 atan2_lookup(f32 y, f32 x) {
  * Compute the angle from (0, 0) to (x, y) as a s16. Given that terrain is in
  * the xz-plane, this is commonly called with (z, x) to get a yaw angle.
  */
-s32 atan2s(f32 y, f32 x) {
+s16 atan2s(f32 y, f32 x) {
     u16 ret;
     if (x >= 0) {
         if (y >= 0) {
